Gans_E-Scooter_Data_Scraper/requirements.txt
python-dotenv
pandas
requests
SQLAlchemy
pymysql